% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/century_check_overwrite.R
\name{century_check_overwrite}
\alias{century_check_overwrite}
\title{Title}
\usage{
century_check_overwrite(file = file, overwrite = TRUE, verbose = TRUE)
}
\arguments{
\item{overwrite}{}
}
\description{
Title
}
\examples{
write.table("Example", "example.txt")
check_overwrite("example.txt")
check_overwrite("example.txt", overwrite = FALSE)
check_overwrite("e.txt")
unlink("example.txt")

# This will generate an error message
write.table("Example", "example.txt")
check_overwrite("example.txt", overwrite = FALSE)
unlink("example.txt")

}
