% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/century_file.R
\name{century_file}
\alias{century_file}
\title{Write *.100 input files for Century}
\usage{
century_file(
  x,
  file = "",
  wd = NULL,
  ndigits = 3,
  overwrite = TRUE,
  sep = "       "
)
}
\arguments{
\item{x}{\code{list} with as many elements as events as needed. Each event starts with a short label and a long title.}

\item{file}{\code{character} string with the name of the file to be created. Only names accepted by CENTURY are accepted.}

\item{wd}{directory to save '*.100' file.}

\item{ndigits}{number of digits for numeric values.}

\item{overwrite}{\code{logical}, if set to TRUE file will be overwritten if it already exists on disk.}
}
\value{
Nothing. File is created on disk.
}
\description{
\code{century_file} allows users to create input files (extension '.100') for the Century soil model
}
\examples{
# Example list with two elements.
set.seed(100)
x <- list()
x[[1]] <- list("E1", "Example1")
x[[1]] <- append(x[[1]], runif(11))
names(x[[1]]) <- c("label", "title", paste0("cultra(", 1:7, ")"), paste0("clteff(", 1:4, ")"))
x[[2]] <- list("E2", "Example2")
x[[2]] <- append(x[[2]], runif(11))
names(x[[2]]) <- c("label", "title", paste0("cultra(", 1:7, ")"), paste0("clteff(", 1:4, ")"))

# Fewer digits.
x[[1]]$`cultra(3)` <- 0.3

# Create file locally.
wd <- ".//data-raw//example"
century_file(x, "cult", wd = wd)
}
