% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/century_list100.R
\name{century_list100}
\alias{century_list100}
\title{Run list100.exe file}
\usage{
century_list100(
  file = file,
  fileout = fileout,
  wd = wd,
  params = params,
  overwrite = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{params}{}
}
\description{
\code{century_list100} runs \code{list100.exe} files
}
\examples{
wd <- ".//data-raw//example"
file_in <- "duke"
file_out <- "out"
params <- list(start_time = 1,
  end_time = 1000,
  args = c("fcacc", "rlvacc", "frtacc", "fbracc", "rlwacc"))
century_list100(file_in, file_out, wd, params)
}
