% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/century_read_lis.R
\name{century_read_lis}
\alias{century_read_lis}
\title{Title}
\usage{
century_read_lis(file = file, wd = wd, verbose = verbose)
}
\arguments{
\item{filename}{}
}
\description{
Title
}
\examples{
wd <- ".//data-raw//example"
params <- list(simulation ="duke",
  output = "out",
  start_time = 1,
  end_time = 100,
  args = c("fcacc", "rlvacc", "frtacc", "fbracc", "rlwacc"))
list100(wd, params)
df <- read_lis(wd, "out")
}
