% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/century_weather.R
\name{century_weather}
\alias{century_weather}
\title{Write CENTURY weather file}
\usage{
century_weather(
  df = df,
  fileout = fileout,
  wd = "",
  overwrite = TRUE,
  sep = "      ",
  ndigits = 3,
  verbose = TRUE
)
}
\arguments{
\item{df}{\code{data.frame} containing weather data (precipitation, maximum and minimum temperature).}

\item{fileout}{\code{character} string with the name (without extension) of the weather file to be created.}

\item{wd}{\code{character} string with the name of a valid directory to save the weather file. Default is the local folder.}

\item{overwrite}{\code{logical} indicating whether to overwrite 'fileout' if it already exists on disk. Default is 'TRUE'.}

\item{sep}{\code{character} string to separate columns by in 'fileout'. Default is an empty \code{character} string of length 7.}

\item{ndigits}{\code{numeric} indicating the number of decimal digits in output file. Default is 3.}

\item{verbose}{\code{logical} indicating whether to print out informative messages. Default is 'TRUE'.}
}
\value{
NULL is returned and fileout is written on disk.
}
\description{
\code{century_weather}
}
\details{
The input \code{data.frame} is pivoted from long to wide format to fit CENTURY requirements.
}
\examples{
df<- data.frame(month = rep(1:12, 2),
  year = rep(c(2020, 2021), each = 12),
  prec = runif(24)*500 + 100,
  tmin = rnorm(24, mean = 5),
  tmax = rnorm(24, mean = 15))
century_weather(df, "nothing")
}
