% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fields.R
\name{check_fields}
\alias{check_fields}
\title{Check correctness of field names for a '.100' data set.}
\usage{
check_fields(x, filename = filename)
}
\arguments{
\item{x}{\code{list} containing the structure of a CENTURY '.100' file}

\item{filename}{}
}
\description{
Check correctness of field names for a '.100' data set.
}
