% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_numeric_character.R
\name{check_numeric_character}
\alias{check_numeric_character}
\title{Check whether input is equal to '?', numeric or a character string.}
\usage{
check_numeric_character(y, type = "min")
}
\arguments{
\item{y}{\code{character} string containing a number, a quotation mark '?' or text.}

\item{type}{\code{character}}
}
\description{
Check whether input is equal to '?', numeric or a character string.
}
\examples{
check_numeric_character("?")
check_numeric_character("?", "max")
check_numeric_character("a")
check_numeric_character("-5")
check_numeric_character("34.3")
}
