% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_overwrite.R
\name{check_overwrite}
\alias{check_overwrite}
\title{Title}
\usage{
check_overwrite(pathname = pathname, filename = filename, overwrite = FALSE)
}
\arguments{
\item{pathname}{\code{character} string with valid path to a folder.}

\item{filename}{\code{character} string specifying the name of the file with its extension.}

\item{overwrite}{\code{logical}, see 'Details'.}
}
\value{
Nothing.
}
\description{
\code{check_overwrite}
}
\examples{
# Create a silly example text file.
sink("example.100")
i <- 1:10
outer(i, i)
sink()

# Checks.
check_overwrite("", "example.txt")
check_overwrite("", "example.txt", overwrite = FALSE)
check_overwrite("", "e.txt")

# Remove the silly example text file.
invisible(file.remove("example.100"))

}
