% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_path.R
\name{check_path}
\alias{check_path}
\title{Check that path exists and that file name, together with extension, has been correctly specified}
\usage{
check_path(path = path, filename = filename, extension = NULL, verbose = TRUE)
}
\arguments{
\item{verbose}{}
}
\description{
Check that path exists and that file name, together with extension, has been correctly specified
}
