% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_path_file.R
\name{check_path_file}
\alias{check_path_file}
\title{Check path and files are correct}
\usage{
check_path_file(pathname = pathname, filename = filename)
}
\arguments{
\item{pathname}{\code{character} string with valid path to a folder.}

\item{filename}{\code{character} string specifying the name of the file with its extension.}
}
\value{
Nothing.
}
\description{
\code{check_path_file} just checks that path and file names are correct, and that the former
corresponds to a valid path.
}
\examples{
# Create a silly example text file.
sink("example.100")
i <- 1:10
outer(i, i)
sink()

check_path_file(getwd(), "example.100")

# Remove the silly example text file.
invisible(file.remove("example.100"))

}
