% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_path_filename.R
\name{check_path_filename}
\alias{check_path_filename}
\title{Title}
\usage{
check_path_filename(path = path, filename = filename, file_site = FALSE)
}
\arguments{
\item{file_site}{}
}
\description{
Title
}
