% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_read.R
\name{check_read}
\alias{check_read}
\title{Check path, file name and extension}
\usage{
check_read(pathname = pathname, filename = filename)
}
\arguments{
\item{pathname}{\code{character} string with valid path to a folder.}

\item{filename}{\code{character} string specifying the name of the file with its extension.}
}
\value{
If successful, it returns nothing.
}
\description{
\code{check_read} verifies that path name exists and that file name, together with extension,
has been correctly specified.
}
\examples{

# Create a silly example text file.
sink("example.100")
i <- 1:10
outer(i, i)
sink()

# Check it exists and it has a valid CENTURY file extension.
check_read(getwd(), "example.100")

# Remove the silly example text file.
invisible(file.remove("example.100"))
}
