% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_write.R
\name{check_write}
\alias{check_write}
\title{Title}
\usage{
check_write(path, filename, overwrite = TRUE)
}
\arguments{
\item{overwrite}{}
}
\value{
Nothing.
}
\description{
Title
}
\details{
\code{check_write} checks that a) path to folder exists and that b) file is removed if 'overwrite' is set to TRUE (default).
This clears the way for functions that write CENTURY files to folders.
}
