% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cult.R
\name{cult}
\alias{cult}
\title{Write 'cult.100' file on disk}
\usage{
cult(x, wd = NULL, ndigits = 3, overwrite = TRUE)
}
\arguments{
\item{x}{\code{list} containing...}

\item{wd}{string with valid file path.}

\item{ndigits}{number of digits with which to round.}
}
\value{
Nothing. A file names 'cult.100' is written on disk.
}
\description{
\code{cult} writes on disk a file named 'cult.100' onto disk.
}
\details{
Straightforward. Default value is zero for all elements.
}
\examples{

# Example list with two elements.
x <- list()
x[[1]] <- list("E1", "Example1")
x[[1]] <- append(x[[1]], runif(11))
names(x[[1]]) <- c("label", "title", paste0("cultra(", 1:7, ")"), paste0("clteff(", 1:4, ")"))
x[[2]] <- list("E1", "Example1")
x[[2]] <- append(x[[2]], runif(11))
names(x[[2]]) <- c("label", "title", paste0("cultra(", 1:7, ")"), paste0("clteff(", 1:4, ")"))

# Fewer digits.
x[[1]]$`cultra(3)` <- 0.3

# Create file locally.
wd <- ".//data-raw//example"
cult(x, wd, ndigits = 2)
}
