% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_100.R
\docType{data}
\name{data_100}
\alias{data_100}
\title{Labels and description for data points in '.100' CENTURY files.}
\format{
A list where each element includes a \code{tibble} with all information about variables.
}
\source{
CENTURY user manual.
}
\usage{
data_100
}
\description{
Labels and description for data points in '.100' CENTURY files.
}
\keyword{datasets}
