% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description_output.R
\docType{data}
\name{description_output}
\alias{description_output}
\title{Description of CENTURY output variables.}
\format{
A \code{data.frame} with 6 columns
}
\source{
CENTURY user manual.
}
\usage{
description_output
}
\description{
Description of all fields that are available as outputs '*.bin' CENTURY files.
}
\keyword{datasets}
