% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_params.R
\name{dirichlet_params}
\alias{dirichlet_params}
\title{Calculate parameters required in a Dirichlet distribution}
\usage{
dirichlet_params(mean = mean, variance = variance)
}
\arguments{
\item{mean}{\code{numerical} vector, same length as \code{variance}, containing the mean values of
component of a Dirichlet distribution.}

\item{variance}{\code{numerical} vector, with only one non-NA element with a \code{variance} value.}
}
\value{
The corresponding \code{alpha} values for the Dirichlet distribution. If the \code{alpha} values are
negative the variance should be decreased. A warning is issued in this case.
}
\description{
\code{dirichlet_params} calculates the 'alpha' values of a Dirichlet distribution from the mean values
and one variance value.
}
\details{
Input \code{mean} vector must contain mean values for each component of the Dirichlet distribution.
If one value is missing (signaled with a NA) it is calculated from the sum of all other mean values.
In turn, only one \code{variance} value must be introduced and all other elements in the \code{variance}
vector must have an \code{NA}.
}
\examples{
# Mean and variance values as input.
m <- c(.3, .23,.27, .2)
v <- c(NA, .1, NA, NA)
alpha <- dirichlet_params(m, v)

# Calculate mean values from alpha.
print(alpha/sum(alpha))

# Calculate variance values from alpha.
alpha0 <- sum(alpha)
beta <- alpha / alpha0
print(beta*(1-beta)/(1+alpha0))

}
