% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_100.R
\docType{data}
\name{files_100}
\alias{files_100}
\title{Number or rows in each CENTURY file with *.100 extension, except for the \if{html}{\out{<site>}}.100 file.}
\format{
A simple named vector.
}
\source{
CENTURY user manual.
}
\usage{
files_100
}
\description{
Number or rows in each CENTURY file with *.100 extension, except for the \if{html}{\out{<site>}}.100 file.
}
\keyword{datasets}
