% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_list.R
\name{fill_list}
\alias{fill_list}
\title{Create a list with simulations}
\usage{
fill_list(df, type = NULL)
}
\arguments{
\item{df}{\code{dataframe} or \code{tibble} with content of file.}

\item{type}{file name the data come from.}
}
\value{
A \code{list}.
}
\description{
Create a list with simulations
}
