% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\name{fire}
\alias{fire}
\title{Schedules a fire in the current month}
\usage{
fire(x, wd = NULL, ndigits = 3, overwrite = TRUE)
}
\arguments{
\item{ndigits}{}
}
\description{
Schedules a fire in the current month
}
