% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_numeric_string.R
\name{is_numeric_string}
\alias{is_numeric_string}
\title{Check if argument is numeric or not.}
\usage{
is_numeric_string(s)
}
\arguments{
\item{a}{\code{vector} object.}
}
\value{
A logical \code{vector}.
}
\description{
Check if argument is numeric or not.
}
