% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels_schedule.R
\docType{data}
\name{labels_schedule}
\alias{labels_schedule}
\title{L
Labels in header to CENTURY schedule file}
\format{
A simple vector with
}
\source{
CENTURY user manual.
}
\usage{
labels_schedule
}
\description{
These are all labels that should be expected in the header of a CENTURY schedule file.
}
\keyword{datasets}
