% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list100.R
\name{list100}
\alias{list100}
\title{Title}
\usage{
list100(wd, params)
}
\arguments{
\item{params}{}
}
\description{
Title
}
\examples{
wd <- ".//data-raw//example"
params <- list(simulation ="duke",
  output = "out",
  start_time = 1,
  end_time = 100,
  args = c("fcacc", "rlvacc", "frtacc", "fbracc", "rlwacc"))
list100(wd, params)
}
