% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_century_files.R
\name{read_century_files}
\alias{read_century_files}
\title{Title}
\usage{
read_century_files(
  path = path,
  filename = filename,
  file_site = FALSE,
  remove_blanks = TRUE
)
}
\arguments{
\item{check_site}{}
}
\description{
Title
}
