% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lis.R
\name{read_lis}
\alias{read_lis}
\title{Title}
\usage{
read_lis(wd, filename)
}
\arguments{
\item{filename}{}
}
\description{
Title
}
\examples{
wd <- ".//data-raw//example"
params <- list(simulation ="duke",
  output = "out",
  start_time = 1,
  end_time = 100,
  args = c("fcacc", "rlvacc", "frtacc", "fbracc", "rlwacc"))
list100(wd, params)
df <- read_lis(wd, "out")
}
