% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_schedule.R
\name{read_schedule}
\alias{read_schedule}
\title{title
Read *.sch CENTURY files.}
\usage{
read_schedule(path, filename)
}
\arguments{
\item{path}{\code{character} string containing a valid path to a *.sch file.}

\item{filename}{\code{character} string with the name of the schedule file. It must include the extension '.sch'.}
}
\value{
\code{list} with two elements named 'header' and 'block_inf', each one being also a \code{list}.
The former list includes basic schedule information that is common to all blocks and that is stored at the
beginning of the file. The latter list has as many elements as treatment blocks, and within each block there are some
info bits specific to that block and a three-column \code{data.frame}, stored in an element named 'Schedule', with the timing
of treatment. Columns are labelled 'year', 'month' and 'option'.
}
\description{
\code{read_schedule} reads *.sch CENTURY soil model files and saves their content in a list.
}
