% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_weather.R
\name{read_weather}
\alias{read_weather}
\title{Read '*.wth' CENTURY files}
\usage{
read_weather(path = path, filename = filename)
}
\arguments{
\item{path}{\code{character} string containing a valid path to a *.sch file.}

\item{filename}{\code{character} string with the name of the weather file.
It must include the extension '.wth'.}
}
\value{
A \code{data.frame} containing the monthly climatic data for the site.
}
\description{
\code{read_weather} reads '*.wth' weather CENTURY files and gives a
wide-format data.frame
}
\details{
The output \code{data.frame} is presented in wide format with 5 columns,
namely 'year', 'month', 'prec', 'tmax' and 'tmin'. Columns are sorted in ascending
order by year and month.

NA values in CENTURY weather files are identified by -99.99 by default.
}
