% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_single_quotes.R
\name{remove_single_quotes}
\alias{remove_single_quotes}
\title{Removal of leading and end single quotation marks.}
\usage{
remove_single_quotes(x)
}
\arguments{
\item{x}{\code{character} vector.}
}
\value{
A \code{character} vector of the same length as 'x' with all leading and end single quotation marks removed.
Input is unchanged if it is numeric.
}
\description{
\code{remove_single_quotes} removes leading and trailing quotes in a character string if present.
}
\examples{
remove_single_quotes("'This leading quote will disappear, but not this 'one.")
}
