% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cent.R
\name{run_cent}
\alias{run_cent}
\title{Run century_47.exe}
\usage{
run_cent(pathname, args)
}
\arguments{
\item{pathname}{\code{character} string with valid path to a folder.}

\item{args}{\code{character} vector with 3 elements, containing the name of the schedule file,
the name of the '.bin' file (without the '.bin' extension) to save the result in, and a
flag "Y" or "N" to decide whether the computation corresponds to an extended one (see CENTURY manual
for details).}
}
\value{
Any message from the execution.
}
\description{
Run century_47.exe
}
\examples{
# See example in \code{\link{century_run}}.
}
