% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_list.R
\name{run_list}
\alias{run_list}
\title{Title}
\usage{
run_list(pathname, name_bin, name_lis, name_txt)
}
\arguments{
\item{pathname}{\code{character} string with valid path to a folder.}

\item{name_bin}{\code{character} string with the name of the '.bin' file to read.}

\item{name_lis}{\code{character} string with the name of the ASCII '.lis' file to create.}

\item{name_txt}{\code{character} string with the name of the ASCII file containing the name of the
fields to extract from 'name_lis'.}
}
\value{
Any message from the execution.
}
\description{
Title
}
\examples{
# See example in \code{\link{century_run}}.
}
