% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitin2.R
\name{splitin2}
\alias{splitin2}
\title{Title}
\usage{
splitin2(x, input_type = input_type, remove_quotes = TRUE)
}
\arguments{
\item{input_type}{element.}

\item{remove_quotes}{}
}
\value{
A \code{data.frame} with one row and two columns.
}
\description{
Title
}
