% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_100.R
\name{write_100}
\alias{write_100}
\title{Write *.100 input files for Century}
\usage{
write_100(
  x,
  pathname = pathname,
  filename = filename,
  ndigits = 6,
  sep = "    ",
  overwrite = FALSE
)
}
\arguments{
\item{x}{\code{list} with as many elements as events as needed. Each event starts with a short label and a long title.}

\item{pathname}{directory to save '*.100' file.}

\item{filename}{description}

\item{ndigits}{number of digits for numeric values.}
}
\value{
Nothing. File is created on disk.
}
\description{
\code{century_file} allows users to create input files (extension '.100') for the Century soil model.
}
\details{
File is overwritten.
}
\examples{
# Example list with two elements.
x <- list()

df <- data.frame(runif(11), c(paste0("cultra(", 1:7, ")"), paste0("clteff(", 1:4, ")")))
colnames(df) <- NULL
x[[1]] <- list(label = "E1", title = "Example1", df = df)

df <- data.frame(runif(11), c(paste0("cultra(", 1:7, ")"), paste0("clteff(", 1:4, ")")))
colnames(df) <- NULL
x[[2]] <- list(label = "E1", title = "Example1", df = df)

# Create temporary path.
path <- tempdir()
write_100(x, path, "cult.100")

}
