% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_schedule.R
\name{write_schedule}
\alias{write_schedule}
\title{Save to disk a *.sch CENTURY file}
\usage{
write_schedule(x, path, filename, overwrite = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{\code{list} with content of the schedule file.}

\item{path}{valid path to write the file to.}

\item{filename}{\code{character} string with the actual file name. It must include the extension '.sch'.}

\item{overwrite}{\code{logical}, if set to TRUE and the file already exists, it will overwrite; if set to
FALSE and the file exists, it will stop with an error message.}

\item{verbose}{\code{logical},}
}
\description{
\code{write_schedule} writes to disk an ASCII *.sch file to be used by the CENTURY soil model.
}
