% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_site.R
\name{write_site}
\alias{write_site}
\title{Title}
\usage{
write_site(
  x,
  pathname = pathname,
  filename = filename,
  ndigits = 6,
  sep = "    ",
  overwrite = FALSE
)
}
\arguments{
\item{overwrite}{}
}
\description{
Title
}
