% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/century_run.R
\name{century_run}
\alias{century_run}
\title{Title}
\usage{
century_run(
  pathname = pathname,
  schedule = schedule,
  name_bin = name_bin,
  name_lis = name_lis,
  name_txt = name_txt,
  overwrite = TRUE,
  extended = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
