% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/century_run.R
\name{century_run}
\alias{century_run}
\title{Title}
\usage{
century_run(
  schedule = schedule,
  fileout = fileout,
  wd = wd,
  extended = FALSE,
  overwrite = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
