% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_100.R
\name{check_100}
\alias{check_100}
\title{Check that path and name of the file with extension *.100 is correct}
\usage{
check_100(path = path, filename = filename)
}
\arguments{
\item{path}{\code{character} string with valid path to a folder.}

\item{filename}{\code{character} string with an actual file name. It must include the extension '.100'.}
}
\description{
Check that path and name of the file with extension *.100 is correct
}
\details{
\code{check_100} does not check that file exists, only that it has a '.100' extension and
that folder exists. If 'check_site' is set to FALSE (default), it also checks that 'filename'
conforms to input file names used by CENTURY.
}
