% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_overwrite.R
\name{check_overwrite}
\alias{check_overwrite}
\title{Title}
\usage{
check_overwrite(path = path, filename = filename, remove = FALSE)
}
\arguments{
\item{overwrite}{}
}
\description{
Title
}
\examples{
write.table("Example", "example.txt")
check_overwrite("example.txt")
check_overwrite("example.txt", overwrite = FALSE)
check_overwrite("e.txt")
unlink("example.txt")

# This will generate an error message
write.table("Example", "example.txt")
check_overwrite("example.txt", overwrite = FALSE)
unlink("example.txt")

}
