% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_path_file.R
\name{check_path_file}
\alias{check_path_file}
\title{Check path and files are correct}
\usage{
check_path_file(pathname = pathname, filename = filename)
}
\arguments{
\item{filename}{\code{character} string with file name.}

\item{path}{\code{character} string with path name.}
}
\value{
Nothing.
}
\description{
\code{check_path_file} just checks that path and file names are correct, and that the former
corresponds to a valid path.
}
