% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_read.R
\name{check_read}
\alias{check_read}
\title{Check path, file name and extension}
\usage{
check_read(path = pathname, filename = filename)
}
\arguments{
\item{path}{\code{character} string indicating the path to the folder.}

\item{filename}{\code{character} string specifying the name of the file with its extension.}
}
\description{
Check that path exists and that file name, together with extension, has been correctly specified
}
