% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_write.R
\name{check_write}
\alias{check_write}
\title{Title}
\usage{
check_write(path = pathname, filename = filename, overwrite = FALSE)
}
\arguments{
\item{filename}{\code{character} string with an actual file name.}

\item{overwrite}{\code{logical}, see 'Details'.}

\item{pathname}{\code{character} string with valid path to a folder.}
}
\value{
Nothing.
}
\description{
Title
}
\details{
\code{check_write} does nothing if 'filename' does not exist on disk.
If it does, then if 'overwrite' is set to TRUE, 'filename' is deleted from disk, and if set to FALSE,
(default) execution stops with a message.
}
