% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_100.R
\name{read_100}
\alias{read_100}
\title{Read *.100 files.}
\usage{
read_100(path, filename, remove_blanks = TRUE)
}
\arguments{
\item{path}{\code{character} string with valid path to write the file to.}

\item{filename}{\code{character} string with the actual file name. It must include the extension '.100'.}

\item{remove_blanks}{\code{logical}, if set to TRUE any blanks at the end of the file will be removed.}
}
\value{
A list with the contents of the file.
}
\description{
\code{read_100} reads almost all CENTURY files with extension '*.100'.
The only *.100 files that cannot be read with \code{read_100} are site files.
}
