% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_100.R
\name{read_100}
\alias{read_100}
\title{Read *.100 files.}
\usage{
read_100(pathname = pathname, filename = filename, remove_blanks = TRUE)
}
\arguments{
\item{pathname}{\code{character} string with valid path to a folder.}

\item{filename}{\code{character} string specifying the name of the file, which must include the extension '.100'.}

\item{remove_blanks}{\code{logical}, if set to TRUE any blanks at the end of the file will be removed.}
}
\value{
A list with the contents of the file.
}
\description{
\code{read_100} reads all CENTURY files with extension '*.100' save for site files, which also have
a '.100' extension but have their own function to read them, called \code{\link{read_site}}.
}
\examples{
# Read one of the CENTURY example files included in the package.
pathname <- system.file("extdata/1.soil_texture_ppt",  package = "Rcentury")
x <- read_100(pathname, "harv.100")

}
