% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitin2.R
\name{splitin2}
\alias{splitin2}
\title{Title}
\usage{
splitin2(
  x,
  input_type = match.arg(input_type, c("data", "title")),
  remove_quotes = TRUE
)
}
\arguments{
\item{remove_quotes}{}
}
\description{
Title
}
