% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_100.R
\name{write_100}
\alias{write_100}
\title{Write *.100 input files for Century}
\usage{
write_100(
  x,
  pathname = pathname,
  filename = filename,
  ndigits = 6,
  sep = "    ",
  overwrite = FALSE
)
}
\arguments{
\item{x}{\code{list} with as many elements as events as needed. Each event starts with a short label and a long title.}

\item{pathname}{\code{character} string with valid path to a folder.}

\item{filename}{\code{character} string specifying the name of the file with its extension.}

\item{ndigits}{\code{integer} specifying the number of digits to print for numeric values.}

\item{sep}{the field separator string. Values within each row of x are separated by this string.}

\item{overwrite}{\code{logical}, if set to FALSE (default), the code will stop if file already exists. If set to
TRUE, the previous file will be overwritten.}
}
\value{
Nothing. File is created on disk.
}
\description{
\code{write_100} allows users to create input files (extension '.100') for the CENTURY soil model.
}
\details{
At this stage there are almost no checks as to whether fields are correctly specified. That's left as a researcher's task.
}
\examples{
# Example list with two elements.
x <- list()

df <- data.frame(runif(11), c(paste0("cultra(", 1:7, ")"), paste0("clteff(", 1:4, ")")))
colnames(df) <- NULL
x[[1]] <- list(label = "E1", title = "Example1", df = df)

df <- data.frame(runif(11), c(paste0("cultra(", 1:7, ")"), paste0("clteff(", 1:4, ")")))
colnames(df) <- NULL
x[[2]] <- list(label = "E1", title = "Example1", df = df)

# Create temporary path.
path <- tempdir()
write_100(x, path, "cult.100")

}
