% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_100.R
\name{write_100}
\alias{write_100}
\title{Write *.100 input files for Century}
\usage{
write_100(
  x,
  path = path,
  filename = filename,
  ndigits = 4,
  sep = "    ",
  verbose = TRUE
)
}
\arguments{
\item{x}{\code{list} with as many elements as events as needed. Each event starts with a short label and a long title.}

\item{path}{directory to save '*.100' file.}

\item{ndigits}{number of digits for numeric values.}
}
\value{
Nothing. File is created on disk.
}
\description{
\code{century_file} allows users to create input files (extension '.100') for the Century soil model.
}
\details{
File is overwritten.
}
\examples{
# Example list with two elements.
set.seed(100)
x <- list()
x[[1]] <- list("E1", "Example1")
x[[1]] <- append(x[[1]], runif(11))
names(x[[1]]) <- c("label", "title", paste0("cultra(", 1:7, ")"), paste0("clteff(", 1:4, ")"))
x[[2]] <- list("E2", "Example2")
x[[2]] <- append(x[[2]], runif(11))
names(x[[2]]) <- c("label", "title", paste0("cultra(", 1:7, ")"), paste0("clteff(", 1:4, ")"))

# Fewer digits.
x[[1]]$`cultra(3)` <- 0.3

# Create file locally.
wd <- ".//data-raw//example"
write_100(x, "cult", wd = wd)

# Read one of the CENTURY example files included in the package.
path <- system.file("extdata/1.soil_texture_ppt",  package = "Rcentury")
x <- read_100(path, "harv.100")


}
